+++ resolwe/flow/serializers/collection.py
@@ -1,11 +1,10 @@
 """Resolwe collection serializer."""
 
 import logging
-from typing import Optional
 
 from rest_framework import serializers
 
-from resolwe.flow.models import Collection, Data, DescriptorSchema
+from resolwe.flow.models import Collection, DescriptorSchema
 from resolwe.rest.fields import ProjectableJSONField
 
 from .base import ResolweBaseSerializer
@@ -20,7 +19,7 @@ class BaseCollectionSerializer(ResolweBaseSerializer):
 
     settings = ProjectableJSONField(required=False)
     data_count = serializers.SerializerMethodField(required=False)
-    status = serializers.SerializerMethodField(required=False)
+    status = serializers.CharField(read_only=True)
 
     def get_data_count(self, collection: Collection) -> int:
         """Return number of data objects on the collection."""
@@ -31,44 +30,6 @@ class BaseCollectionSerializer(ResolweBaseSerializer):
             if hasattr(collection, "data_count")
             else collection.data.count()
         )
-
-    def get_status(self, collection: Collection) -> Optional[str]:
-        """Return status of the collection based on the status of data objects.
-
-        When collection contains no data objects None is returned.
-        """
-
-        status_order = [
-            Data.STATUS_ERROR,
-            Data.STATUS_UPLOADING,
-            Data.STATUS_PROCESSING,
-            Data.STATUS_PREPARING,
-            Data.STATUS_WAITING,
-            Data.STATUS_RESOLVING,
-            Data.STATUS_DONE,
-        ]
-
-        # Use 'data_statuses' attribute when available. It is created in the
-        # BaseCollectionViewSet class. It contains all the distinct statuses of the
-        # data objects in the collection.
-        status_set = (
-            set(collection.data_statuses)
-            if hasattr(collection, "data_statuses")
-            else collection.data.values_list("status", flat=True).distinct()
-        )
-
-        if not status_set:
-            return None
-
-        for status in status_order:
-            if status in status_set:
-                return status
-
-        logger.warning(
-            "Could not determine the status of a collection.",
-            extra={"collection": collection.__dict__},
-        )
-        return None
 
     class Meta:
         """CollectionSerializer Meta options."""